+++ CHANGELOG.md
@@ -9,10 +9,11 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v5.47.1
+
 ### Fixed
 
 - `RelationDirective`/`batchload_relations` will consider `null` and default connection name the same connection https://github.com/nuwave/lighthouse/pull/2125
-
 
 ## v5.47.0